# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SmartProBonoAPP API",
    "description": "API documentation for the SmartProBonoAPP legal services platform",
    "version": "1.0.0"
  },
  "host": "localhost:5003",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
    "/document-generator/templates": {
      "get": {
        "summary": "Get document templates",
        "description": "Returns a list of available document templates",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Template"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/document-generator/generate": {
      "post": {
        "summary": "Generate document",
        "description": "Generates a document based on the provided template and data",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Document generation request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DocumentGenerationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/DocumentGenerationResponse"
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticates a user and returns an access token",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/cases": {
      "get": {
        "summary": "Get user cases",
        "description": "Returns a list of cases associated with the authenticated user",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Case"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {
    "Template": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "required_fields": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "DocumentGenerationRequest": {
      "type": "object",
      "properties": {
        "templateId": {
          "type": "integer",
          "format": "int64"
        },
        "data": {
          "type": "object"
        }
      },
      "required": ["templateId", "data"]
    },
    "DocumentGenerationResponse": {
      "type": "object",
      "properties": {
        "documentId": {
          "type": "string"
        },
        "downloadUrl": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["email", "password"]
    },
    "LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "Case": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "client_name": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": ["NEW", "IN_PROGRESS", "UNDER_REVIEW", "ON_HOLD", "CLOSED"]
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
} 